rgb2hsv_add_a_channel

+++ rgb2hsv.h
@@ -19,14 +19,14 @@ struct rgb {
     double r;       // a fraction between 0 and 1
     double g;       // a fraction between 0 and 1
     double b;       // a fraction between 0 and 1
-    unsigned int a;
+    double a;       // a fraction between 0 and 1
 };
 
 struct hsv {
     double h;       // angle in degrees
     double s;       // a fraction between 0 and 1
     double v;       // a fraction between 0 and 1
-    unsigned int a;
+    double a;       // a fraction between 0 and 1
 };
 
 static struct hsv   rgb2hsv(struct rgb in);
